style: apply eslint auto fix to entire project
- ESLint --fix 옵션을 통한 프로젝트 전반의 코드 스타일 자동 교정 반영
- 들여쓰기, 세미콜론, 사용하지 않는 변수 정리 등 코드 컨벤션 최적화

+++ apps/api/src/modules/affiliate/referral/application/link-referral.service.ts
@@ -30,7 +30,7 @@ export class LinkReferralService {
     private readonly findCodeByCodeService: FindCodeByCodeService,
     private readonly policy: ReferralPolicy,
     private readonly dispatchLogService: DispatchLogService,
-  ) { }
+  ) {}
 
   @Transactional()
   async execute({
@@ -66,7 +66,7 @@ export class LinkReferralService {
     // 4. 레퍼럴 관계 생성
     const referral = await this.repository.create({
       affiliateId: code.userId,
-      codeId: code.id as bigint,
+      codeId: code.id,
       subUserId,
       ipAddress,
       deviceFingerprint,
